app/requirements.txt
Flask
kafka-python
pandas
pyarrow